requirements.txt
altgraph==0.17.4
asgiref==3.8.1
blinker==1.9.0
click==8.1.8
colorama==0.4.6
dj-database-url==2.3.0
Django==5.1.7
et_xmlfile==2.0.0
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
openpyxl==3.1.5
packaging==24.2
pefile==2023.2.7
psycopg==3.2.6
psycopg-binary==3.2.6
psycopg2-binary==2.9.10
pywin32-ctypes==0.2.3
setuptools==78.1.0
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
Werkzeug==3.1.3
whitenoise==6.9.0
Pillow>=10.0.0
django-ckeditor>=6.7.0
django-ratelimit>=4.1.0
bleach>=6.1.0
redis>=5.0.0
